refactor: use `field_serializer` for `CONNECTOR_SCOPE` for consistency
