+++ connectors-sdk/connectors_sdk/settings/base_settings.py
@@ -129,12 +129,27 @@ class _BaseConnectorConfig(BaseConfigModel, ABC):
         description="The name of the connector.",
     )
     scope: ListFromString = Field(
-        description="The scope of the connector, e.g. 'flashpoint'."
+        description="The scope of the connector, e.g. 'indicator, vulnerability'."
     )
     log_level: Literal["debug", "info", "warn", "warning", "error"] = Field(
         description="The minimum level of logs to display.",
         default="error",
     )
+
+    @field_serializer("scope", mode="wrap", when_used="json")
+    def _serialize_scope(
+        self,
+        value: Any,
+        handler: SerializerFunctionWrapHandler,
+        info: FieldSerializationInfo,
+    ) -> str | list[str]:
+        """Serialize scope as a comma-separated string when serializing for `pycti.OpenCTIConnectorHelper` only.
+        Otherwise, return the list of strings.
+        """
+        mode = info.context.get("mode") if info.context else None
+        if isinstance(value, list) and mode == "pycti":
+            return ",".join(value)  # [ "e1", "e2", "e3" ] -> "e1,e2,e3"
+        return handler(value)  # type: ignore[no-any-return] # actually return `list[str]`
 
 
 class BaseExternalImportConnectorConfig(_BaseConnectorConfig):
