+++ connectors-sdk/tests/data/.env.test
@@ -2,6 +2,6 @@ OPENCTI_URL=http://localhost:8080
 OPENCTI_TOKEN=changeme
 CONNECTOR_ID=connector-poc--uid
 CONNECTOR_NAME=Test Connector
-CONNECTOR_SCOPE=test
+CONNECTOR_SCOPE=scope1,scope2
 CONNECTOR_LOG_LEVEL=debug
 CONNECTOR_DURATION_PERIOD=PT5M
